refactor: Re-init with uniorg next-blog-starter
`yarn create next-app --example https://github.com/rasendubi/uniorg/tree/master/examples/next-blog-starter blog-starter-app`

Making a design decision for ease of use. Going to have one code base
for the blog(this one), and a seperate code base to host a digital
garden. I wanted everything to live in the digitial garden, but I think
I can handle the few links pointing into it, and let the few people that
read it get lost in the garden from there. Main thing is I'll be able to
write in org and not markdown. I'd like this to be in svelte, but I
think I need to build dogleg first and better understand svelte, and
right now the more important thing is being able to make content,
without giving up future flexibility that I want. All of this for
org-mode.